# internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: |
    API для аутентификации пользователей (выдача, обновление, отзыв токенов, получение GUID).
servers:
  - url: /api/v1
paths:
  /auth/tokens:
    post:
      operationId: IssueTokens
      summary: Выдать новую пару токенов для пользователя
      description: |
        Возвращает новую пару access/refresh токенов для пользователя с указанным GUID. Требует API ключ.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: guid
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Пара токенов выдана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Ошибка аутентификации (неверный API ключ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/tokens/refresh:
    post:
      operationId: RefreshTokens
      summary: Обновить пару токенов
      security:
        - BearerAuth: []
      description: |
        Обновляет access/refresh токены. Принимает refresh-токен из http-only cookie, а access-токен из заголовка Authorization.
      responses:
        '200':
          description: Пара токенов обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokensResponse'
        '401':
          description: Ошибка аутентификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      operationId: Logout
      summary: Деавторизация пользователя
      description: |
        Удаляет все refresh-сессии пользователя (отзыв токенов).
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Успешно
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/user/guid:
    get:
      operationId: GetUserGUID
      summary: Получить GUID текущего пользователя
      description: |
        Возвращает GUID пользователя по access-токену (Bearer).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: GUID пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGUIDResponse'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TokensResponse:
      type: object
      properties:
        access_token:
          type: string
      required:
        - access_token

    UserGUIDResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
      required:
        - user_id

    ErrorResponse:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason